requirements.txt
asgiref==3.5.2
attrs==22.1.0
aws-cdk.cloud-assembly-schema==1.178.0
aws-cdk.core==1.178.0
aws-cdk.cx-api==1.178.0
aws-cdk.region-info==1.178.0
cattrs==22.2.0
constructs==3.4.133
Django==4.1.1
django-ckeditor==6.5.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
exceptiongroup==1.0.0rc9
jsii==1.70.0
Pillow==9.2.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
typeguard==2.13.3
typing-extensions==4.4.0
tzdata==2022.2
